nimble/porting: Make os_mbuf/os_mpool work on 64-bit.
Pointers are stored in the os_mempool::mp_membuf_addr field, so it needs to be uintptr_t, not uint32_t. Also needs correct alignment.

This was previously fixed in c5a0c7a3ee1718f2c702317e7ec2f3890d5237a3 but ab53154c1cadba70394e76e90eb37c20f2a4a525 changed it back to match mynewt-core.

+++ porting/nimble/src/os_mbuf.c
@@ -244,7 +244,7 @@ os_mbuf_get(struct os_mbuf_pool *omp, uint16_t leadingspace)
 {
     struct os_mbuf *om;
 
-    os_trace_api_u32x2(OS_TRACE_ID_MBUF_GET, (uint32_t)omp,
+    os_trace_api_u32x2(OS_TRACE_ID_MBUF_GET, (uint32_t)(uintptr_t)omp,
                        (uint32_t)(uintptr_t)leadingspace);
 
     if (leadingspace > omp->omp_databuf_len) {
